fix : update table transaction

+++ src/Ressources/Transaction.sql
@@ -1,18 +1,18 @@
 -- create table transaction
 CREATE TABLE IF NOT EXISTS transaction (
     transactionId VARCHAR(255) PRIMARY KEY,
-    amount NUMERIC,
+    amount int,
     description VARCHAR(255),
     accountId VARCHAR(255) REFERENCES account(accountId)
 );
 
 -- insert for transaction
 INSERT INTO transaction (transactionId, amount, description, accountId)
-VALUES ('txn001', 50.00, 'Purchase of goods', '123456')
+VALUES ('txn001', 50, 'Purchase of goods', '123456')
 ON CONFLICT (transactionId) DO NOTHING;
 
 INSERT INTO transaction (transactionId, amount, description, accountId)
-VALUES ('txn002', -20.00, 'Refund for returned item', '789012')
+VALUES ('txn002', -20, 'Refund for returned item', '789012')
 ON CONFLICT (transactionId) DO NOTHING;
 
 INSERT INTO transaction (transactionId, amount, description, accountId)
